♻️ refactor: separate TaskEditor class into taskEditor.js for better code structure
- Moved TaskEditor class to taskEditor.js to improve code modularity
- Enhanced maintainability by keeping related functionality in a dedicated file
- Simplified main script by separating concerns
