+++ src/board/handlers/taskHandlers.js
@@ -7,7 +7,7 @@ import {
   getISOStringNow,
   detectDeviceType,
 } from '../../shared/utils/common.js';
-import { createEditForm } from '../renderers/template.js';
+import TaskEditor from '../renderers/taskEditor.js';
 
 // 새 카드 생성
 function createNewTask(target) {
@@ -103,64 +103,6 @@ function toggleModifyFrom(target) {
   const targetCard = target.closest('.task-item');
   const editor = new TaskEditor(targetCard);
   editor.showEditForm();
-}
-
-class TaskEditor {
-  constructor(taskCard) {
-    this.taskCard = taskCard;
-    this.originalTaskCard = taskCard.cloneNode(true);
-    this.editForm = this.createEditForm();
-  }
-
-  createEditForm() {
-    const title = this.taskCard.querySelector('.task-title').textContent;
-    const content = this.taskCard.querySelector('.task-content').textContent;
-    const editFormElement = createEditForm(title, content);
-
-    this.attachFormEvents(editFormElement);
-
-    return editFormElement;
-  }
-
-  attachFormEvents(editFormElement) {
-    editFormElement
-      .querySelector('.task-cancel-btn')
-      .addEventListener('click', () => {
-        this.restoreOriginalTask();
-      });
-
-    editFormElement
-      .querySelector('.task-save-btn')
-      .addEventListener('click', () => {
-        this.saveEdit();
-      });
-  }
-
-  restoreOriginalTask() {
-    this.editForm.replaceWith(this.originalTaskCard);
-  }
-
-  saveEdit() {
-    const editedTitle = this.editForm.querySelector('input').value;
-    const editedContent = this.editForm.querySelector('textarea').value;
-
-    this.originalTaskCard.querySelector('.task-title').textContent =
-      editedTitle;
-    this.originalTaskCard.querySelector('.task-content').textContent =
-      editedContent;
-
-    updateTask({
-      id: this.originalTaskCard.id,
-      title: editedTitle,
-      content: editedContent,
-    });
-
-    this.restoreOriginalTask();
-  }
-
-  showEditForm() {
-    this.taskCard.replaceWith(this.editForm);
-  }
 }
 
 export {
